requirements.txt
pytest==2.9.0
pytest_bdd
selenium
jsonpickle
pony
PyMySQL
mysql-connector-python-rf
lxml-4.6.3-cp39-cp39-win32.whl
allure-pytest